site/requirements.txt
django~=4.0
django-crispy-forms~=1.12
crispy-bootstrap5~=0.3
psycopg2-binary~=2.8
libsass~=0.20
django-sass-processor~=1.0
markdown~=3.3
bleach~=3.3
factory-boy~=3.2
django-autoslug~=1.9
isort~=5.10
black==21.12b0
flake8~=4.0.1
flake8-django~=1.1
flake8-black~=0.2
flake8-isort~=4.1
django-tree-queries==0.6.0
Pygments~=2.10
Pillow~=8.3
sheatless~=1.0.1
PyPDF2~=1.26.0
django-ical~=1.8.0
python-magic~=0.4.24
